Chore: Created `autoprefixer` gulp task runner
1) npm install gulp-autoprefixer
2) var autoprefixer = require('gulp-autoprefixer');

When Sass encounters an error, it just stops. We most likely don't want this behavior. We want Sass to finish and just log the error. We do this by adding .on('error', sass.logError)) after .pipe(sass)

So the complete new code for the styles task will be:

For Gulp 3:

gulp.task('styles', function() {
    gulp.src('sass/**/*.scss')
        .pipe(sass()).on('error', sass.logError)
        .pipe(gulp.dest('./css'));
});

Since Autoprefixer works on CSS files, we can use it in the same task as our Sass. After we run sass, we will run autoprefixer. We also tell autoprefixer that we want it to work with the last 2 browser versions, though we can change this if we like.

Our new styles task with Autoprefixer:

For Gulp 3:

gulp.task('styles', function() {
    gulp.src('sass/**/*.scss')
        .pipe(sass()).on('error', sass.logError)
        .pipe(autoprefixer({
            browsers: ['last 2 versions']
        }))
        .pipe(gulp.dest('./css'));
});